+++ src/components/Home/Home.css
@@ -1,5 +1,5 @@
 
-.container-landing{
+.container-landing-home{
     display: flex;
     justify-content: center;
     align-items: center;
