+++ src/components/Home/Home.jsx
@@ -1,47 +1,20 @@
 import React from 'react'
-import Grid from '@mui/material/Grid';
-import Console from '../../Assets/Console.svg'
-import Box from '@mui/material/Box';
-import { styled } from '@mui/material/styles';
-import Header from '../Header/Header';
-
-import './Home.css'
-import { Typography } from '@mui/material';
+import Land from '../Pages/Land';
+import Show from '../Pages/Show';
+import End from '../Pages/End';
 
 
-const Root = styled('img')(({ theme }) => ({
-    padding: theme.spacing(1),
-    [theme.breakpoints.down('md')]: {
-        width: 360
-      
-    },
-    [theme.breakpoints.up('md')]: {
-        width: 533
-
-    },
-    [theme.breakpoints.up('lg')]: {
-        width: 533
-
-    },
-  }));
 
 
 
 export default function Home() {
   return (
-      <Box sx={{ flexGrow: 1,  backgroundColor: '#9c2a2a', height: '100vh', justifyContent: 'center', alignItems: 'center'}}>
-          <Header /> 
-    <Grid container className='container-landing'>
-             <Typography>
-    <h1> Get New and trending</h1>
-   <h4>Based on player counts and release date</h4> 
-    </Typography>
-        <Grid item xs={12}>
-            <Root src={Console} className='img-game' />
-        </Grid>
-        
-        </Grid>
-      </Box>
+      <>
+      <Land />
+      <Show />
+      <End />
+      </>
+
 
   )
 }
