+++ src/components/Pages/End.css
@@ -1,0 +1,12 @@
+.container-landing{
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    padding: 40px 0;
+
+}
+
+
+.img-game{
+    width: 633px
+}
