+++ src/components/Pages/Land.css
@@ -1,0 +1,11 @@
+
+.container-land{
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    padding: 40px 0;
+    flex-direction: column;
+
+}
+
+
